難易度選択完了時のイベントを別ファイルに切り出した
